+++ src/main/java/com/tagnumelite/chickens/api/recipe/IBreedingRecipe.java
@@ -1,0 +1,10 @@
+package com.tagnumelite.chickens.api.recipe;
+
+import net.minecraft.world.Container;
+import net.minecraft.world.item.ItemStack;
+
+public interface IBreedingRecipe extends ISpecialRecipe<Container> {
+    ItemStack getFirstParent();
+
+    ItemStack getSecondParent();
+}