+++ src/main/java/com/tagnumelite/chickens/api/recipe/IThrowingRecipe.java
@@ -1,0 +1,6 @@
+package com.tagnumelite.chickens.api.recipe;
+
+import net.minecraft.world.Container;
+
+public interface IThrowingRecipe extends ISpecialRecipe<Container> {
+}