+++ src/main/java/com/tagnumelite/chickens/crafting/recipes/ThrowingRecipe.java
@@ -1,0 +1,41 @@
+package com.tagnumelite.chickens.crafting.recipes;
+
+import com.tagnumelite.chickens.api.recipe.IThrowingRecipe;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.Container;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.RecipeSerializer;
+import net.minecraft.world.item.crafting.RecipeType;
+import net.minecraft.world.level.Level;
+
+public class ThrowingRecipe implements IThrowingRecipe {
+    @Override
+    public boolean matches(Container pContainer, Level pLevel) {
+        return false;
+    }
+
+    @Override
+    public ItemStack assemble(Container pContainer) {
+        return null;
+    }
+
+    @Override
+    public ItemStack getResultItem() {
+        return null;
+    }
+
+    @Override
+    public ResourceLocation getId() {
+        return null;
+    }
+
+    @Override
+    public RecipeSerializer<?> getSerializer() {
+        return null;
+    }
+
+    @Override
+    public RecipeType<?> getType() {
+        return null;
+    }
+}